Only emit ``lru-cache-decorating-method`` when ``maxsize`` is ``None`` (#6181)
