+++ doc/whatsnew/2.13.rst
@@ -78,11 +78,12 @@ New checkers
    enter on a non specialized keyboard.
    See `Confusable Characters in PEP 672 <https://www.python.org/dev/peps/pep-0672/#confusable-characters-in-identifiers>`_
 
-* Added ``lru-cache-decorating-method`` checker with checks for the use of ``functools.lru_cache``
-  on class methods. This is unrecommended as it creates memory leaks by never letting the instance
-  getting garbage collected.
+* Added ``cache-max-size-none`` checker with checks for the use of ``functools.lru_cache``
+  on class methods with a ``maxsize`` of ``None``. This is unrecommended as it
+  creates memory leaks by never letting the instance get garbage collected.
 
   Closes #5670
+  Clsoes #6180
 
 Removed checkers
 ================
