--- tests/functional/l/lru_cache_decorating_method.txt
@@ -1,12 +1,0 @@
-lru-cache-decorating-method:27:5:27:14:MyClassWithMethods.my_func:'lru_cache' without 'maxsize' will keep all method args alive indefinitely, including 'self':INFERENCE
-lru-cache-decorating-method:31:5:31:24:MyClassWithMethods.my_func:'lru_cache' without 'maxsize' will keep all method args alive indefinitely, including 'self':INFERENCE
-lru-cache-decorating-method:35:5:35:32:MyClassWithMethods.my_func:'lru_cache' without 'maxsize' will keep all method args alive indefinitely, including 'self':INFERENCE
-lru-cache-decorating-method:39:5:39:18:MyClassWithMethods.my_func:'lru_cache' without 'maxsize' will keep all method args alive indefinitely, including 'self':INFERENCE
-lru-cache-decorating-method:43:5:43:16:MyClassWithMethods.my_func:'lru_cache' without 'maxsize' will keep all method args alive indefinitely, including 'self':INFERENCE
-lru-cache-decorating-method:47:5:47:26:MyClassWithMethods.my_func:'lru_cache' without 'maxsize' will keep all method args alive indefinitely, including 'self':INFERENCE
-lru-cache-decorating-method:51:5:51:34:MyClassWithMethods.my_func:'lru_cache' without 'maxsize' will keep all method args alive indefinitely, including 'self':INFERENCE
-lru-cache-decorating-method:55:5:55:20:MyClassWithMethods.my_func:'lru_cache' without 'maxsize' will keep all method args alive indefinitely, including 'self':INFERENCE
-lru-cache-decorating-method:60:5:60:20:MyClassWithMethods.my_func:'lru_cache' without 'maxsize' will keep all method args alive indefinitely, including 'self':INFERENCE
-lru-cache-decorating-method:61:5:61:20:MyClassWithMethods.my_func:'lru_cache' without 'maxsize' will keep all method args alive indefinitely, including 'self':INFERENCE
-lru-cache-decorating-method:75:5:75:26:MyClassWithMethodsAndMaxSize.my_func:'lru_cache' without 'maxsize' will keep all method args alive indefinitely, including 'self':INFERENCE
-lru-cache-decorating-method:79:5:79:26:MyClassWithMethodsAndMaxSize.my_func:'lru_cache' without 'maxsize' will keep all method args alive indefinitely, including 'self':INFERENCE
